refactor: consolidate sessions/ into session/
Move sessions/ root files (MeepleSessionCard, SessionQuotaBar) and
sessions/live/ subdirectory into session/. Update all imports across
6 page files. The old sessions/ directory is removed.
