+++ apps/web/src/components/session/index.ts
@@ -1,4 +1,14 @@
-// Session Toolkit Components - Barrel Export
+// Session Components - Consolidated Barrel Export
+
+// Session cards (formerly components/sessions/)
+export { SessionQuotaBar, type SessionQuotaBarProps } from './SessionQuotaBar';
+export {
+  MeepleSessionCard,
+  MeepleSessionCardSkeleton,
+  type MeepleSessionCardProps,
+} from './MeepleSessionCard';
+
+// Session Toolkit Components
 
 export { LiveIndicator } from './LiveIndicator';
 export { LiveScoreSheet } from './LiveScoreSheet';
