+++ apps/web/src/components/session/live/GuestScoreProposal.tsx
