+++ apps/web/src/components/session/live/ScoreBoard.tsx
