+++ apps/web/src/components/session/live/SessionNavConfig.tsx
