+++ apps/web/src/components/session/live/__tests__/PausedSessionCard.test.tsx
