+++ apps/web/src/components/session/live/__tests__/ScoreBoard.test.tsx
