--- apps/web/src/components/sessions/index.ts
@@ -1,10 +1,0 @@
-/**
- * Sessions Components (Issue #3075)
- *
- * Components related to game session management and quota display.
- */
-
-export { SessionQuotaBar, type SessionQuotaBarProps } from './SessionQuotaBar';
-export { MeepleSessionCard, MeepleSessionCardSkeleton, type MeepleSessionCardProps } from './MeepleSessionCard';
-
-// Note: Storybook stories are in SessionQuotaBar.stories.tsx
